Make possible to just allocate a callback in a pool
The use case for this was to set already allocated strings (maybe
statically or somewhere else) to also be freed with a callback. Not the
actual string structure because we don't know where that may be, but the
internal storage that we may have malloc'd.

This required adding closures to pool callbacks. I added some comments
on improvements we could have on the pool API.