Don't str_set() an uninitialized string_t
Now that I removed all calls of ZERO_INIT() on string_t, I still had
memory issues because str_set() was using uninitialized values. When a
new string_t is created we DO want to call ZERO_INIT() on it, or assign
the result of calling new_str(). I removed the static assert on
ZERO_INIT().

The idea about making calls of ZERO_INIT() on string_t a compile error
sounded like a good one. But it defeats the purpose supporting a zero
initialized string_t to mean an empty string.

The real problem here is that there is a distinction between
initializing a string_t for the first time, and resetting an already
initialized string to be empty. Calling ZERO_INIT() when it's actually
the second case, leaks memory. A better solution to the problem would be
to make easy to track when this incorrect assignments happen. Maybe add
a leak flag that marks a striing as soon as we get a call to
str_maybe_grow() this means that the string is being used, then we can
use gdb to break if that flag ever gets cleared?. Maybe find a way not
to require gdb here.