Add a common_unused.h file where unused but working functions will go
I want common.h to be as small and useful as possible, if I see
something isn't useful in general I migh put it in this file and remove
it from common.h. Still this code should work, in case it proves to be
useful and we want it back.