Fix BINARY_TREE_FOR() for real
The problem wasn't the initialization, traversal to the leftmost node
has to be done for each iteration, the condition inside the reference
implementation avoided visiting these nodes implicitly. The translation
to a for loop missed this detail.

Added a visit_node variable that marks when to visit a node and when to
skip it.