Don't hide a call to sh_expand() inside file handling functions
This will cause breakage downstream, but I really think sh_expand()
should be called explicitly if necessary, because it causes problems
when a file path contains () which is very common for example in names
of files downloaded multiple times. Really, an alternative way of
getting absolute paths should be used, shell expansion is problematic.