Remove special parsing for dates without space and UTC offset
I had a special case to support parsing dates like 1900-02:30 where the
date is just a year and it has a UTC offset of -02:30. Here we needed to
backtrack because initially the 02 would be interpreted as a month until
we fount the ':' character. I now think this was not a good idea as it
complicates parsing and supports a format that is hard to parse visually
by humans. Instead, these dates are now invalid, to represent such date
we now require a space (or T), like '1900 -02:30'.