Implement computation of week of date and its generalization
Further development of recurrent events implementation. Improved test
cases and added a function that will allow computing if a date is part
of a recurrent event in constant time, based on Zeller's congruence
normally used to compute the day of the week of a date.

Added a test case that uses this new constant time congruence
computation together with the day addition funtionality. It checks that
both produce consisten results with eachother.