Improvements to test_logger.c
Implement mechanism for sending a string from a subprocess, and show it
in the test output even when the child process exists successfully. Add
an internal string_t whose content is transfered through a temporary
file, into the parent process.

Add mechanism to allow changing the string that represents a successful
test. Currently used to replace the string in negative tests from "OK"
to "-OK".