Fix computation of no-option arguments
The only way to do this correctly is to keep track of which properties
have been queried and if they had an argument or not. The API breaks but
most likely any usage of the old one was broken anyway.

Keeping track of the context is optional so that we don't add overhead
in the simpler case where an executable does not use together properties
and the no-option arguments.