Correctly write date when available and requested precisions don't match
Sometimes we bay fix a precision for writing dates (let's say we choose
maximum precision, seconds), still it's possible a date has missing data
in which case it will have fewer available precision fields. In this
case we were failing to add the date/time separator when writing a date.

Fix the condition that writes the separator and add new tests for this
issue.