Remove _tree prefix from binary tree's generated code
This was a huge mental overhead because it wasn't consistent, some
functions had the prefix, others didn't. Now we just don't have it. This
is much easier to remember as we just need to learn a single word for
generated types and functions.

Tree type only has _t appended, _node_t for nodes, _insert(), _lookup(),
_get(), etc.