Additions to the string library
Added starts_with(), str_rstrip(), cstr_find_open_parenthesis() and
cstr_find_close_parenthesis(). The later ones are useful when supporting
processing of strings where we want to find balanced parenthesis.

Fixed an issue in linked list sorting where we were not auto computing
the list's size to be 0 because of an early return.

Add remove last API for the dynamic array API.

Reqoek the test logger's shorthand functions to use the new API that
allows afoiding the closing test_pop() call and instead uses a single
condition so specify the error and success branches.